requirements.txt
pandas==1.1.4
scikit_optimize==0.8.1
seaborn==0.11.0
tensorflow==2.3.1
matplotlib==3.3.2
numpy==1.18.5
PyYAML==5.3.1
scikit_learn==0.23.2

